backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
celery>=5.3
redis>=5.0
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
boto3>=1.28
pytesseract>=0.3
Pillow>=10.0
python-dotenv>=1.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.2.0
pandas>=2.0.0
opencv-python>=4.8.0 